backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
google-generativeai==0.8.3
google-auth==2.49.1
pydantic==2.10.4
python-multipart==0.0.20
python-dotenv==1.0.1
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
requests==2.33.1
